Keenan Branch 1 Merge into Keenan Branch Card Gameplay (#2)
* Camera moved and scale test models

Camera has been moved for down view. models add for scale testing.

* place holders added

place holder for the playing field have been added

* SceneOrganisation

* Keenan: Moved the scene to the KeenansScenes folder

* TCP

* Waterstuff

* Renamed some files

* Testing

* akjedgf

* WaterShader

a nice water shader for people to use in their environments

* Added models

land and plant added

* Update LevelBockOut.unity

* Keenan: Created a scene

Created a scene where the user can chose whether to start a client or a host (client and server). Based off of Matt's TCP server connectivity code.

* Keenan: Add board state script

* Keenan: Added placeholder cards to the scene

* Keenan: Made the host device automatically connect

Made the host device automatically connect the internal client to the server.

* Keenan: Added feature

Added the ability for the user of the client to enter the IP address of the host they want to connect to.

* Keenan: Rearranged UI to be more intuitive

---------

Co-authored-by: bengreen210 <ben.green210@gmail.com>
Co-authored-by: Finleyfooonfire <finleyfooonfire@gmail.com>
Co-authored-by: mattez1 <133816825+mattez1@users.noreply.github.com>